Move „Added to Cart“ message to component
